+++ test/wButton.test.js
@@ -21,7 +21,24 @@ describe('button', () => {
   test('primary', () => {
     const wrapper = mount(wButton, { props: { primary: true, label } })
     assert.equal(wrapper.text(), 'Hello Warp')
-    assert.include(wrapper.classes().join(' '), ccButton.buttonPrimary)
+    assert.include(wrapper.classes().join(' '), ccButton.primary)
+  })
+  test('negative', () => {
+    const wrapper = mount(wButton, { props: { negative: true, label } })
+    assert.equal(wrapper.text(), 'Hello Warp')
+    assert.include(wrapper.classes().join(' '), ccButton.negative)
+  })
+  test('no variant should default to secondary', () => {
+    const wrapper = mount(wButton)
+    assert.include(wrapper.classes().join(' '), ccButton.secondary)
+  })
+  test('small with no variant should default to secondary small', () => {
+    const wrapper = mount(wButton, { props: { small: true } })
+    assert.include(wrapper.classes().join(' '), ccButton.secondarySmall)
+  })
+  test('small & quiet with no variant should default to secondary small quiet', () => {
+    const wrapper = mount(wButton, { props: { small: true, quiet: true } })
+    assert.include(wrapper.classes().join(' '), ccButton.secondarySmallQuiet)
   })
   test('href', () => {
     const href = 'https://finn.no'